refactor: extract duplicate generateKenBurnsConfig to shared utility
Move randomInRange and generateKenBurnsConfig from BoothPage.tsx and Slideshow.tsx

into src/utils/kenBurns.ts to eliminate duplication and reduce maintenance risk.

Closes #128
